refactor(#164): annotations/pmi.py — PMI annotation pass (P5c)
Move _annotate_pmi (the PMI/GD&T annotation pass) into annotations/pmi.py. It is
self-contained — its placement helpers (_try_above/_try_below/_try_left/_try_right,
_witness_from_bbox, _bore_info) are nested inside it, and it uses no shared
annotations helpers, only _core._dim and dwg methods — so it lifts cleanly with no
new _common dependency. (The annotation pass; the STEP-side PMI extraction stays in
draftwright.pmi.)

annotate.py re-imports _annotate_pmi (called by _auto_annotate); pmi.py imports
only _core + build123d/features, never annotate/make_drawing — no cycle. Pure move:
golden gate passes against the committed snapshots with no regeneration; PMI tests
pass; ruff check + format + mypy clean. Adversarially checked: no upward import, no
stale def.

Part of #164 (P5d holes + P5e orchestrator follow).
